chore: remove projectfilereader class
